Update index.html
age should be a dropdown to simplify for parents. second an state change that deactivates the button and shows a loading symbol until and answer is returned. third a short paragraph above the first form field briefly (as possible) describing to the user what they will see back when info is entered to the form